# docs/swagger.yaml
basePath: /
definitions:
...
  handler.AssignAirdropRequest:
    properties:
      airdrop_id:
        example: 1
        type: integer
      custom_tasks:
        items:
          $ref: '#/definitions/handler.CustomTaskRequest'
        type: array
      notes:
        example: Focus on bridging
        type: string
      template:
        description: basic, full, custom
        example: basic
        type: string
    type: object
...
  handler.CloneAccountRequest:
    properties:
      color:
        example: '#EF4444'
        type: string
      name:
        example: Akun 2
        type: string
    type: object
  handler.ComparisonRow:
    properties:
      account_color:
        type: string
      account_id:
        type: integer
      account_name:
        type: string
      completed_airdrops:
        type: integer
      completed_tasks:
        type: integer
      pending_tasks:
        type: integer
      total_airdrops:
        type: integer
      total_tasks:
        type: integer
      wallet_count:
        type: integer
    type: object
  handler.CreateAccountRequest:
    properties:
      color:
        example: '#3B82F6'
        type: string
      name:
        example: Akun 1
        type: string
      notes:
        example: Main sybil account
        type: string
    type: object
...
  handler.CustomTaskRequest:
    properties:
      description:
        example: Bridge 0.1 ETH
        type: string
      frequency:
        example: weekly
        type: string
    type: object
...
  handler.UpdateAccountRequest:
    properties:
      color:
        example: '#EF4444'
        type: string
      name:
        example: Akun 1
        type: string
      notes:
        example: Updated notes
        type: string
    type: object
  model.Account:
    properties:
      account_airdrops:
        items:
          $ref: '#/definitions/model.AccountAirdrop'
        type: array
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      wallets:
        description: Relations
        items:
          $ref: '#/definitions/model.Wallet'
        type: array
    type: object
  model.AccountAirdrop:
    properties:
      account:
        allOf:
        - $ref: '#/definitions/model.Account'
        description: Relations
      account_id:
        type: integer
      airdrop:
        $ref: '#/definitions/model.Airdrop'
      airdrop_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      notes:
        type: string
      status:
        description: active, completed, paused
        type: string
      tasks:
        items:
          $ref: '#/definitions/model.Task'
        type: array
      updated_at:
        type: string
    type: object
  model.Airdrop:
    properties:
      category:
        type: string
      chain:
        type: string
      created_at:
        type: string
      deadline:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      priority:
        type: string
      status:
        type: string
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: integer
    type: object
  model.Task:
    properties:
      account_airdrop:
        allOf:
        - $ref: '#/definitions/model.AccountAirdrop'
        description: Relations
      account_airdrop_id:
        type: integer
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      frequency:
        description: once, daily, weekly, monthly
        type: string
      gas_spent:
        type: number
      id:
        type: integer
      is_completed:
        type: boolean
      next_due:
        type: string
      tx_hash:
        type: string
    type: object
...
  model.Wallet:
    properties:
      account:
        allOf:
        - $ref: '#/definitions/model.Account'
        description: Relations
      account_id:
        type: integer
      address:
        type: string
      chain:
        type: string
      created_at:
        type: string
      id:
        type: integer
      label:
        type: string
      user_id:
        type: integer
    type: object
host: airdrop-tracker-api-v1.fly.dev
info:
  contact:
    email: support@airdrop-tracker.local
    name: Bita
  description: Personal airdrop task management API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Airdrop Tracker API
  version: "1.0"
paths:
...
  /api/accounts:
    get:
      description: Get all sybil accounts for authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Account'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all accounts
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Create new sybil account
      parameters:
      - description: Account data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create account
      tags:
      - Accounts
  /api/accounts/{id}:
    delete:
      description: Delete account (only if empty, or use ?force=true)
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Force delete with all data
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - Accounts
    get:
      description: Get single account by ID with wallets and account-airdrops
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Account'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account detail
      tags:
      - Accounts
    put:
      consumes:
      - application/json
      description: Update account name, color, notes
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update account
      tags:
      - Accounts
  /api/accounts/{id}/airdrops:
    get:
      description: Get all airdrop assignments for a specific account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AccountAirdrop'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List airdrops for account
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: Assign a global airdrop to an account with optional task template
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Assignment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AssignAirdropRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.AccountAirdrop'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign airdrop to account
      tags:
      - Accounts
  /api/accounts/{id}/airdrops/{airdrop_id}:
    delete:
      description: Unlink an airdrop from an account and delete associated tasks
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Airdrop ID
        in: path
        name: airdrop_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove airdrop from account
      tags:
      - Accounts
  /api/accounts/{id}/clone:
    post:
      consumes:
      - application/json
      description: Clone an account with its airdrop assignments and tasks (not wallets)
      parameters:
      - description: Source Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: New account data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CloneAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Clone account
      tags:
      - Accounts
...
  /api/dashboard/comparison:
    get:
      description: Get per-account progress stats for comparison
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.ComparisonRow'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get comparison table
      tags:
      - Dashboard
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"